Better seperation between admin and user accessible pages
