+++ app/views/admin/space_types/edit.html.erb
