+++ app/views/admin/space_types/new.html.erb
