--- app/views/space_types/_form.html.erb
@@ -1,4 +1,0 @@
-<%= simple_form_for @space_type do |f| %>
-  <%= f.input :type_name %>
-  <%= f.submit %>
-<% end %>
