--- app/views/space_types/index.html.erb
@@ -1,11 +1,0 @@
-<div>
-  <% @space_types.each do |space_type| %>
-    <div>
-      <%= link_to 'Edit', edit_space_type_path(space_type) %>
-      <%= link_to 'Delete', space_type, method: :delete, data: { confirm: 'Are you sure?' } %>
-      <%= space_type.type_name %>
-    </div>
-  <% end %>
-
-  <%= link_to 'New Space Type', new_space_type_path %>
-</div>
