+++ app/views/spaces/show.html.erb
@@ -54,7 +54,7 @@
       <section class="bg-pink-50 shadow-lg p-4 rounded-lg border border-lnu-pink">
         <h3 class="no-mt pt-1">Admin:</h3>
         <p>
-          <%= link_to t("space_show.show_changelog"), admin_path(@space) %>
+          <%= link_to t("space_show.show_changelog"), admin_history_path(@space) %>
         </p>
       </section>
     <% end %>
